requirements.txt
alembic==1.7.5
anyio==3.3.4
asgiref==3.4.1
autopep8==1.5.5
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
blinker==1.4
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
debugpy==1.5.0
decorator==5.1.0
distlib==0.3.3
entrypoints==0.3
Faker==9.9.0
fastapi==0.70.0
filelock==3.1.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.3
idna==3.3
ijson==3.1.4
importlib-metadata==4.8.2
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
json-tools==0.4.1
jsondiff==1.3.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
nest-asyncio==1.5.1
numpy==1.21.2
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
pika==1.1.0
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.4.0
prettytable==0.7.2
prompt-toolkit==3.0.20
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3

pyzmq==22.3.0
requests==2.22.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
spark-parser==1.8.9
SQLAlchemy==1.4.26
sqlparse==0.4.2
starlette==0.16.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
uncompyle6==3.8.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
xdis==6.0.3
zipp==3.6.0
